requirements.txt
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.1.7
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
lxml==5.3.1
oscrypto==1.3.0
pdfminer.six==20250327
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pyparsing==3.2.3
pypdf==5.4.0
python-bidi==0.6.6
python-docx==1.1.2
python-dotenv==1.1.0
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.1
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
xhtml2pdf==0.2.17
